CodeGen from PR 28271 in Azure/azure-rest-api-specs Merge b6ad3dd4a2696511a436abcacd81a064dae042f4 into 0a008454a630680b7e8e688d89665248a0e62ac0

+++ sdk/avs/azure-resourcemanager-avs/src/main/java/com/azure/resourcemanager/avs/models/WorkloadNetworkDhcpServer.java
@@ -9,7 +9,9 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 
-/** NSX DHCP Server. */
+/**
+ * NSX DHCP Server.
+ */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "dhcpType")
 @JsonTypeName("SERVER")
 @Fluent
@@ -24,15 +26,17 @@ public final class WorkloadNetworkDhcpServer extends WorkloadNetworkDhcpEntity {
      * DHCP Server Lease Time.
      */
     @JsonProperty(value = "leaseTime")
-    private Long leaseTime;
+    private Integer leaseTime;
 
-    /** Creates an instance of WorkloadNetworkDhcpServer class. */
+    /**
+     * Creates an instance of WorkloadNetworkDhcpServer class.
+     */
     public WorkloadNetworkDhcpServer() {
     }
 
     /**
      * Get the serverAddress property: DHCP Server Address.
-     *
+     * 
      * @return the serverAddress value.
      */
     public String serverAddress() {
@@ -41,7 +45,7 @@ public final class WorkloadNetworkDhcpServer extends WorkloadNetworkDhcpEntity {
 
     /**
      * Set the serverAddress property: DHCP Server Address.
-     *
+     * 
      * @param serverAddress the serverAddress value to set.
      * @return the WorkloadNetworkDhcpServer object itself.
      */
@@ -52,32 +56,36 @@ public final class WorkloadNetworkDhcpServer extends WorkloadNetworkDhcpEntity {
 
     /**
      * Get the leaseTime property: DHCP Server Lease Time.
-     *
+     * 
      * @return the leaseTime value.
      */
-    public Long leaseTime() {
+    public Integer leaseTime() {
         return this.leaseTime;
     }
 
     /**
      * Set the leaseTime property: DHCP Server Lease Time.
-     *
+     * 
      * @param leaseTime the leaseTime value to set.
      * @return the WorkloadNetworkDhcpServer object itself.
      */
-    public WorkloadNetworkDhcpServer withLeaseTime(Long leaseTime) {
+    public WorkloadNetworkDhcpServer withLeaseTime(Integer leaseTime) {
         this.leaseTime = leaseTime;
         return this;
     }
 
-    /** {@inheritDoc} */
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public WorkloadNetworkDhcpServer withDisplayName(String displayName) {
         super.withDisplayName(displayName);
         return this;
     }
 
-    /** {@inheritDoc} */
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public WorkloadNetworkDhcpServer withRevision(Long revision) {
         super.withRevision(revision);
@@ -86,7 +94,7 @@ public final class WorkloadNetworkDhcpServer extends WorkloadNetworkDhcpEntity {
 
     /**
      * Validates the instance.
-     *
+     * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     @Override
